# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lilypad API",
        "description": "API for lilypad",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/currency": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "summary": "Get currencies or filter by name",
                "description": "Retrieves all currencies if no query parameters or JSON body is provided. Filters by name using either query parameters or JSON body.",
                "operationId": "9c04f635bfdcfe8f485fe511ce772fa5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the currency to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "USD"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the currency to filter by",
                                        "type": "string",
                                        "example": "USD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currency retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Currency object or list of currencies",
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Currency retrieved successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Currency not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Currency not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Currency"
                ],
                "summary": "Create a new currency",
                "description": "Creates a new currency with details such as name, code, active flag, and currency rates.",
                "operationId": "4bab1d22863d62e27b4898f5a89491d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the currency",
                                        "type": "string",
                                        "example": "US Dollar"
                                    },
                                    "code": {
                                        "description": "Currency code",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "active": {
                                        "description": "Active status of the currency",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "quickBookCurrencyRate": {
                                        "description": "QuickBooks currency rate",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1
                                    },
                                    "globalCurrencyRate": {
                                        "description": "Global currency rate",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Currency Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Currency Created Successfully!"
                                        },
                                        "data": {
                                            "description": "Created currency object",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}